LayoutTests/crypto/subtle/ecdh-generate-key-p384-expected.txt

Test generating an EC key pair with P-384 using ECDH algorithm.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Generating a key pair...
PASS keyPair.toString() is '[object Object]'
PASS keyPair.publicKey.type is 'public'
PASS keyPair.publicKey.extractable is true
PASS keyPair.publicKey.algorithm.name is 'ECDH'
PASS keyPair.publicKey.algorithm.namedCurve is 'P-384'
PASS keyPair.publicKey.usages is [ ]
PASS keyPair.privateKey.type is 'private'
PASS keyPair.privateKey.extractable is false
PASS keyPair.privateKey.algorithm.name is 'ECDH'
PASS keyPair.privateKey.algorithm.namedCurve is 'P-384'
PASS keyPair.privateKey.usages is ['deriveBits', 'deriveKey']
PASS successfullyParsed is true

TEST COMPLETE
